Mandrill attachments are NOT always Base64 encoded
According to Mandrill's documentation, the attachments array's contents
are NOT always Base64 encoded:
http://help.mandrill.com/entries/22092308-What-is-the-format-of-inbound-email-webhooks-

In particular, text attachments are not Base64 encoded unless they
contain non-UTF-8 characters. By always trying to Base64 decode the
attachments this adapter was corrupting Mandrill text file attachments.

This PR simply respects the base64 flag that Mandrill sends along with
each attachment to know whether or not to decode the content field.